requirements-cloud.txt
# Cloud backend dependencies (no sounddevice/PySide6/webrtcvad)
fastapi
uvicorn[standard]
websockets
requests
numpy
fugashi
unidic-lite
pykakasi
